+++ packages/ui-kit-js/src/components/navigation/link/link.tsx
@@ -1,9 +1,10 @@
 import { Component, Element, h, JSX, Prop } from '@stencil/core';
 import cx from 'classnames';
-import { BreakpointCustomizable, hasShadowDom, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
+import { BreakpointCustomizable, mapBreakpointPropToPrefixedClasses, prefix } from '../../../utils';
 import { IconName } from '../../icon/icon/icon-name';
 import { improveFocusHandlingForCustomElement } from '../../../utils/focusHandling';
 import { LinkTarget, Theme } from '../../../types';
+import { improveLinkHandlingForCustomElement } from '../../../utils/linkHandling';
 
 @Component({
   tag: 'p-link',
@@ -42,28 +43,7 @@ export class Link {
 
   public componentDidLoad() {
     improveFocusHandlingForCustomElement(this.element);
-
-    /**
-     * IE11/Edge (not chromium based) workaround to
-     * fix the event target of click events (which normally
-     * shadow dom takes care of)
-     *
-     * caution: if the click event would be bound on
-     * safari browsers with force touch, the force touch
-     * functionality wouldn't work anymore. but since they
-     * do have shadowDom the click event should never be
-     * bound.
-     */
-    if (!hasShadowDom(this.element)) {
-      this.element.addEventListener('click', (event: MouseEvent): void => {
-        /**
-         * caution: do not preventDefault here, else ie11 wouldn't work
-         * anymore
-         */
-        event.stopPropagation();
-        this.element.click();
-      }, true);
-    }
+    improveLinkHandlingForCustomElement(this.element);
   }
 
   public render(): JSX.Element {
@@ -81,7 +61,12 @@ export class Link {
     return (
       <TagType
         class={linkClasses}
-        {...(TagType === 'a' ? { href: this.href, target: `${this.target}`, download: this.download, rel: this.rel } : null)}
+        {...(TagType === 'a' ? {
+          href: this.href,
+          target: `${this.target}`,
+          download: this.download,
+          rel: this.rel
+        } : null)}
       >
         <p-icon
           class={iconClasses}