+++ packages/ui-kit-js/CHANGELOG.md
@@ -4,9 +4,17 @@
 
 ### Change
 
-* Link
-  * Remove tabbable attribute 
+* Button component
+  * New custom element
+  
+* Button-Pure component
+  * New custom element
+ 
+* Remove button-regular and button-icon component
 
+* Link component
+  * New custom element
+  
 * Text component
   * Reduce amount of text variants in global interface
   * Remove `ul | ol | li | sup | sub | label` from `tag` property